requirements.txt
# Base packages para extração e manipulação de dados
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
pandas-datareader>=0.10.0
yfinance>=0.2.18

# Utilitários e ferramentas
python-dotenv>=1.0.0
loguru>=0.7.0
typer>=0.7.0
requests>=2.28.2
cloudpickle>=2.2.1
tqdm>=4.65.0

# Manipulação de datas e calendário
holidays>=0.25

# Azure Storage
azure-storage-blob>=12.16.0

# Development tools
ipython>=8.0.0